libmythtv: Start to simplify and generalise GPU interop
- firstly, the existing code is very OpenGL centric, so needs to be
generalised to support at least Vulkan for the future
- secondly, with the evolution of the interop and player code, we can
now simply check for the list of supported interops at the start of
playback and pass this to the interop/decoder classes as needed. This
avoids potentially expensive callbacks to check support - and avoids
stale data where the renderer has changed and we were caching supported
types to avoid those expensive callbacks.